requirements.txt
psychopy==2025.1.1
numpy
pandas
pyserial
